Cleanup and additional comments.
