+++ java/lrgs-web/src/main/java/org/opendcs/lrgs/http/LrgsWebApp.java
@@ -10,5 +10,5 @@ public class LrgsWebApp extends Application
 {
     @Context
     ServletContext servletContext;
-    
+
 }
