+++ java/lrgs-web/src/main/java/org/opendcs/lrgs/http/dds/UseDdsSession.java
@@ -6,6 +6,11 @@ import java.lang.annotation.RetentionPolicy;
 /**
  * For resources that will need access to a valid DDS Session to retrieve
  * data from the archive.
+ *
+ * Any jakarta.ws.rs Resource that will interact with the Message Archive
+ * should be marked with this annotation to ensure an {@link lrgs.common.DcpMsgRetriever}
+ * instance avaiable.
+ *
  * UseDdsSession
  */
 @Retention(RetentionPolicy.RUNTIME)
